Update allure helper Remove TagAllureReporter Add OnSchedulerCallback
