+++ pom.xml
@@ -76,7 +76,7 @@
     <dependency>
       <groupId>ru.sbtqa.tag</groupId>
       <artifactId>allure-helper</artifactId>
-      <version>1.1.0</version>
+      <version>1.2.0</version>
       <type>jar</type>
     </dependency>
     <dependency>
